Former-commit-id: 6c2a56c7a0eaf8b1c0ff4263248511948e12389c

Original commit: elastic/x-pack@3c5bfd2b5fb1d2574fad56600f347d797c97e2d2

+++ java/src/test/com/prelert/job/warnings/DummyStatusReporter.java
@@ -39,7 +39,7 @@ public class DummyStatusReporter extends StatusReporter
 	}
 
 	@Override
-	protected void reportStatus(int totalRecords)
+	protected void reportStatus(long totalRecords)
 	{
 		// do nothing
 	}
